Updated permission checks

+++ src/commands/Fun/screenshot.js
@@ -19,6 +19,12 @@ module.exports = class Screenshot extends Command {
 
 	// Run command
 	async run(bot, message, args, settings) {
+		// Make sure bot has permissions to send attachments
+		if (!message.channel.permissionsFor(bot.user).has('ATTACH_FILES')) {
+			bot.logger.error(`Missing permission: \`ATTACH_FILES\` in [${message.guild.id}].`);
+			return message.error(settings.Language, 'MISSING_PERMISSION', 'ATTACH_FILES').then(m => m.delete({ timeout: 10000 }));
+		}
+
 		// make sure a website was entered
 		if (!args[0]) {
 			if (message.deletable) message.delete();
